Minor cleanup and restructure
Little more cleanup

Fixed operator check for adding entity to entityset
